feat: introduce admin settings to left bottom pane
