+++ app/client/cypress/support/AdminSettingsCommands.js
@@ -47,10 +47,8 @@ Cypress.Commands.add("fillGithubForm", () => {
 
 // open authentication page
 Cypress.Commands.add("openAuthentication", () => {
-  cy.get(".t--profile-menu-icon").should("be.visible");
-  cy.get(".t--profile-menu-icon").click();
-  cy.get(".t--admin-settings-menu").should("be.visible");
-  cy.get(".t--admin-settings-menu").click();
+  cy.get(".admin-settings-menu-option").should("be.visible");
+  cy.get(".admin-settings-menu-option").click();
   cy.url().should("contain", "/settings/general");
   // click authentication tab
   cy.get(adminSettings.authenticationTab).click();
