+++ app/client/src/pages/common/ProfileDropdown.tsx
@@ -17,19 +17,10 @@ import {
 import { ReduxActionTypes } from "@appsmith/constants/ReduxActionConstants";
 import ProfileImage from "./ProfileImage";
 import { PopperModifiers } from "@blueprintjs/core";
-import {
-  PROFILE,
-  ADMIN_SETTINGS_CATEGORY_DEFAULT_PATH,
-} from "constants/routes";
+import { PROFILE } from "constants/routes";
 import { Colors } from "constants/Colors";
-import {
-  ACCOUNT_TOOLTIP,
-  createMessage,
-  ADMIN_SETTINGS,
-} from "@appsmith/constants/messages";
+import { ACCOUNT_TOOLTIP, createMessage } from "@appsmith/constants/messages";
 import { TOOLTIP_HOVER_ON_DELAY } from "constants/AppConstants";
-import { useSelector } from "react-redux";
-import { getCurrentUser } from "selectors/usersSelectors";
 
 type TagProps = CommonComponentProps & {
   onClick?: (text: string) => void;
@@ -96,7 +87,6 @@ const UserNameWrapper = styled.div`
 `;
 
 export default function ProfileDropdown(props: TagProps) {
-  const user = useSelector(getCurrentUser);
   const Profile = (
     <TooltipComponent
       content={createMessage(ACCOUNT_TOOLTIP)}
@@ -146,18 +136,6 @@ export default function ProfileDropdown(props: TagProps) {
         }}
         text="Edit Profile"
       />
-      {user?.isSuperUser && user?.isConfigurable && (
-        <StyledMenuItem
-          className={`t--admin-settings-menu ${BlueprintClasses.POPOVER_DISMISS}`}
-          icon="setting"
-          onSelect={() => {
-            getOnSelectAction(DropdownOnSelectActions.REDIRECT, {
-              path: ADMIN_SETTINGS_CATEGORY_DEFAULT_PATH,
-            });
-          }}
-          text={createMessage(ADMIN_SETTINGS)}
-        />
-      )}
       <StyledMenuItem
         className="t--logout-icon"
         icon="logout"
